Updated model.py with indexing problem
A problem occurs when the loop_index + batch_size > minimum number of elements in dataset A/B (min(len(random_order_A), len(random_order_B))). The size of indexes_B is then smaller then batch_size which results in an error because indexes_A and indexes_B will not be of the same size anymore. 

This commit fixes this error.