Updated indexing problem
The previous commit did not work properly. Please take it back as it was before or accept this change. I debugged it and noticed that min_nr_imgs was not correct. The change actually kept taking the same images from dataset A until dataset B was done - which is not good. 

It was correct what you did Simon, but you just missed something for the last iteration of the epoch. I added another if statement to check for the last iteration. I tried running an example with len(A_train) > len(B_train) and reverse and it works as it should.

My suggestion is to either take it back as it was or accept this change.